==> prior_experiments/experiments_501/0.prepare_dataset.ipynb <==
import os
import numpy as np
import pandas as pd

df = pd.read_csv("../prior_experiments/evals_share/courtlistener_dataset_with_overrulings.csv", sep="\t")
len(df)

df.columns

df.head()

file_mapping = {}

for index, row in df.iterrows():
    citing_id = row["citing_opinion_id"]
    cited_id = row["cited_opinion_id"]

    filename = f"{index + 1:04d}.{citing_id}_cites_{cited_id}.txt"
    file_mapping[(citing_id, cited_id)] = filename
    print(f"Processed: {filename}")

# Add filenames to the DataFrame
df["filename"] = df.apply(
    lambda row: file_mapping.get((row["citing_opinion_id"], row["cited_opinion_id"])), axis=1
)

df.columns

df = df[["filename", "cluster_id", "citing_decision_name", "cited_opinion_id", "cited_decision_name", "overruled"]]
df = df.rename(columns={'cluster_id': 'citing_cluster_id', 'cited_opinion_id': 'cited_cluster_id'})

df.head()

citing_df = df[["citing_cluster_id", "citing_decision_name"]].drop_duplicates()
len(citing_df)

cited_df = df[["cited_cluster_id", "cited_decision_name"]].drop_duplicates()
len(cited_df)

df.to_csv("data/dataset.csv", index=False)
citing_df.to_csv("data/citing_dataset.csv", index=False)
cited_df.to_csv("data/cited_dataset.csv", index=False)
