==> prior_experiments/experiments_501/2.create_cleaned_dataset.ipynb <==
import pandas as pd

df = pd.read_csv("data/dataset.csv")
citing = pd.read_json("data/citing_opinions_metadata.json")
cited = pd.read_json("data/cited_opinions_metadata.json")

df.head()

citing.head()

notes = {
'0138.103012_cites_93638.txt': "cited opinion has multiple matches",
'0111.101913_cites_87768.txt': "missing eyecite link in opinion",
'0660.108990_cites_8989511.txt': "cited opinion has multiple matches",
'0034.98124_cites_94455.txt': "missing eyecite link in opinion",
'0243.103442_cites_99906.txt': "cited opinion has multiple matches",
'0414.105319_cites_101915.txt': "cited opinion has multiple matches",
'0429.105525_cites_105413.txt': "missing eyecite link in opinion",
'0631.108350_cites_288328.txt': "cited opinion has multiple matches",
'1037.118011_cites_658639.txt': "cited opinion has multiple matches",
'0050.99004_cites_3485674.txt': "cited opinion has multiple matches",
'0245.103442_cites_97451.txt': "cited opinion has multiple matches",
'1092.118273_cites_746267.txt': "cited opinion has multiple matches",
'0410.104965_cites_104628.txt': "cited opinion has multiple matches",
'0849.111218_cites_409686.txt': "cited opinion has multiple matches",
'0051.99004_cites_93379.txt': "cited opinion has multiple matches",
'0077.101750_cites_101176.txt': "missing eyecite link in opinion",
'1043.118133_cites_438842.txt': "cited opinion has multiple matches",
'0926.112301_cites_507765.txt': "cited opinion has multiple matches",
'0090.101913_cites_101894.txt': "cited opinion has multiple matches",
'0661.108990_cites_8988675.txt': "cited opinion has multiple matches",
'1058.118149_cites_109716.txt': "cited opinion has multiple matches",
'0687.109514_cites_321071.txt': "cited opinion has multiple matches",
'0653.108990_cites_307980.txt': "cited opinion has multiple matches"}

df['note'] = df['filename'].map(notes)

notes_df = pd.DataFrame([
    {
        "note": v,
        "citing_cluster_id": int(k.split(".")[1].split("_")[0]),
        "cited_cluster_id": int(k.split("_")[-1].replace(".txt", ""))
    }
    for k, v in notes.items()
])

notes_df.head()

df = df.merge(
    notes_df,
    on=['citing_cluster_id', 'cited_cluster_id'],
    how='left',
    suffixes=('', '_note')
)

df['note'] = df['note'].combine_first(df['note_note'])
df.drop(columns='note_note', inplace=True)

df['use_full_opinion'] = df['note'].notna().astype(int)
df.head()

df['use_full_opinion'].value_counts()

result = df.merge(citing[["citing_cluster_id", "citing_url", "citing_opinions", "citing_filenames"]], how="left", on="citing_cluster_id")
result.head()

cited.head()

cited.columns

result = result.merge(cited[["cited_cluster_id", "cited_url", 'cited_name_short', 'cited_name', 'cited_name_full', 'cited_citations','cited_opinions', 'cited_filenames']], how="left", on="cited_cluster_id")
result.head()

assert len(df) == len(result)

result.columns

result = result[['filename', 
                 'citing_cluster_id', 'citing_decision_name', 'citing_url', 'citing_opinions', 'citing_filenames', 
                 'cited_cluster_id', 'cited_decision_name', 'cited_url', 'cited_name_short', 'cited_name', 'cited_name_full', 'cited_citations', 
                 'overruled', 'note', 'use_full_opinion',]]
result.head()

result.to_csv("data/cleaned_dataset.csv", index=False)
