==> prior_experiments/experiments_501/3.annotate_opinions.ipynb <==
# Annotate opinions (with full case citation & reference citation) & extract excerpts for all opinions

Use helper functions to annotate opinions and then extract excerpts surrounding the annotations.

# Import libraries

import pandas as pd

from utils.excerpts_utils import process_dataframe

# Load the dataset & call the helper function

df = pd.read_csv("data/cleaned_dataset.csv")
df.head()

%%time

df = process_dataframe(df)
df.head()
