==> prior_experiments/experiments_501/4.combine_excerpts_v1.ipynb <==
# Combined the excerpts to preprare for inference

# Import libraries

import pandas as pd

from utils.combine_utils import combine_excerpts

# Load the data & combine excerpts

df = pd.read_csv("data/cleaned_dataset.csv")
df.head()

df = combine_excerpts(df)
df

df[df["filename"] == "0050.99004_cites_3485674.txt"]

# Check for any empty excerpts

import os

# Set the folder path
folder_path = "data/excerpts/"

# List all txt files in the folder
txt_files = [f for f in os.listdir(folder_path) if f.endswith('.txt')]

# Check each file for being blank
blank_files = []
for file in txt_files:
    file_path = os.path.join(folder_path, file)
    if os.path.getsize(file_path) == 0:  # Checks if file size is 0 bytes
        blank_files.append(file)

# Print the result
if blank_files:
    print("Blank txt files found:", blank_files)
else:
    print("No blank txt files found.")

['0750.109450_cites_108444.txt', ---> wrong cluster id
'0138.103012_cites_93638.txt',   ---> cited opinion has multiple matches
'0111.101913_cites_87768.txt',   ---> missing eyecite link in opinion
'0660.108990_cites_8989511.txt', ---> cited opinion has multiple matches
'0611.108164_cites_97237.txt',   ---> wrong cluster id
'0034.98124_cites_94455.txt',    ---> missing eyecite link in opinion
'0614.108164_cites_268703.txt',  ---> wrong cluster id
'0243.103442_cites_99906.txt',   ---> cited opinion has multiple matches
'0999.117947_cites_663717.txt',  ---> cited cluster id is not part of the citing opinion, replace with a different cluster id
'1064.118230_cites_1087822.txt', ---> wrong cluster id
'0414.105319_cites_101915.txt',  ---> cited opinion has multiple matches
'0429.105525_cites_105413.txt',  ---> missing eyecite link in opinion
'0631.108350_cites_288328.txt',  ---> cited opinion has multiple matches
'1037.118011_cites_658639.txt',  ---> cited opinion has multiple matches
'0050.99004_cites_3485674.txt',  ---> cited opinion has multiple matches
'0245.103442_cites_97451.txt',   ---> cited opinion has multiple matches
'0055.99004_cites_97966.txt',    ---> wrong cluster id
'1092.118273_cites_746267.txt',  ---> cited opinion has multiple matches
'0656.108990_cites_305771.txt',  ---> wrong cluster id
'0513.107564_cites_2443377.txt', ---> wrong cluster id
'0410.104965_cites_104628.txt',  ---> cited opinion has multiple matches
'0849.111218_cites_409686.txt',  ---> cited opinion has multiple matches
'0051.99004_cites_93379.txt',    ---> cited opinion has multiple matches
'0236.103442_cites_2516349.txt', ---> wrong cluster id
'0077.101750_cites_101176.txt',  ---> missing eyecite link in opinion
'0906.111948_cites_108280.txt',  ---> wrong cluster id
'1043.118133_cites_438842.txt',  ---> cited opinion has multiple matches
'0926.112301_cites_507765.txt',  ---> cited opinion has multiple matches
'0090.101913_cites_101894.txt',  ---> cited opinion has multiple matches
'0351.104380_cites_1116658.txt', ---> wrong cluster id
'0661.108990_cites_8988675.txt', ---> cited opinion has multiple matches
'0588.107973_cites_104816.txt',  ---> wrong cluster id
'1058.118149_cites_109716.txt',  ---> cited opinion has multiple matches
'0687.109514_cites_321071.txt',  ---> cited opinion has multiple matches
'0653.108990_cites_307980.txt',  ---> cited opinion has multiple matches]

replaces = {108444: 8985546,
            97237: 5726506,
            268703: 8887067,
            663717: 1879640,
            1087822: 103707,
            97966: 8180960,
            305771: 8901897,
            2443377: 103664,
            2516349: 95716,
            108280: 8982256,
            1116658: 104182,
            104816: 8928994,
           }
            

notes = {'0138.103012_cites_93638.txt': "cited opinion has multiple matches",
'0111.101913_cites_87768.txt': "missing eyecite link in opinion",
'0660.108990_cites_8989511.txt': "cited opinion has multiple matches",
'0034.98124_cites_94455.txt': "missing eyecite link in opinion",
'0243.103442_cites_99906.txt': "cited opinion has multiple matches",
'0414.105319_cites_101915.txt': "cited opinion has multiple matches",
'0429.105525_cites_105413.txt': "missing eyecite link in opinion",
'0631.108350_cites_288328.txt': "cited opinion has multiple matches",
'1037.118011_cites_658639.txt': "cited opinion has multiple matches",
'0050.99004_cites_3485674.txt': "cited opinion has multiple matches",
'0245.103442_cites_97451.txt': "cited opinion has multiple matches",
'1092.118273_cites_746267.txt': "cited opinion has multiple matches",
'0410.104965_cites_104628.txt': "cited opinion has multiple matches",
'0849.111218_cites_409686.txt': "cited opinion has multiple matches",
'0051.99004_cites_93379.txt': "cited opinion has multiple matches",
'0077.101750_cites_101176.txt': "missing eyecite link in opinion",
'1043.118133_cites_438842.txt': "cited opinion has multiple matches",
'0926.112301_cites_507765.txt': "cited opinion has multiple matches",
'0090.101913_cites_101894.txt': "cited opinion has multiple matches",
'0661.108990_cites_8988675.txt': "cited opinion has multiple matches",
'1058.118149_cites_109716.txt': "cited opinion has multiple matches",
'0687.109514_cites_321071.txt': "cited opinion has multiple matches",
'0653.108990_cites_307980.txt': "cited opinion has multiple matches"}

# Save the file for inference

#df.to_csv("data/dataset_output.csv", index=False)
